# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /users/set_password:
    post:
      tags:
      - Пользователи
      summary: Обновление пароля
      operationId: setPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPasswordDto'
        required: true
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /register:
    post:
      tags:
      - Регистрация
      summary: Регистрация пользователя
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
  /login:
    post:
      tags:
      - Авторизация
      summary: Авторизация пользователя
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
...
  /users/me:
    get:
      tags:
      - Пользователи
      summary: Получить информацию об авторизованном пользователе
      operationId: getUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
    patch:
      tags:
      - Пользователи
      summary: Обновление информации об авторизованном пользователе
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserDto'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /users/me/image:
    patch:
      tags:
      - Пользователи
      summary: Обновить аватар авторизованного пользователя
      operationId: updateUserImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
...
  /users/image/{id}/from-db:
    get:
      tags:
      - user-controller
      operationId: getUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: array
                    items:
                      type: string
                      format: byte
...
components:
  schemas:
    NewPasswordDto:
      type: object
      properties:
        currentPassword:
          maxLength: 16
          minLength: 8
          type: string
          description: текущий пароль
        newPassword:
          maxLength: 16
          minLength: 8
          type: string
          description: новый пароль
    RegisterDto:
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
          description: логин
        password:
          maxLength: 16
          minLength: 8
          type: string
          description: пароль
        firstName:
          maxLength: 16
          minLength: 2
          type: string
          description: имя пользователя
        lastName:
          maxLength: 16
          minLength: 2
          type: string
          description: фамилия пользователя
        phone:
          pattern: "\\+7\\s?\\(?\\d{3}\\)?\\s?\\d{3}-?\\d{2}-?\\d{2}"
          type: string
          description: телефон пользователя
        role:
          type: string
          description: роль пользователя
          enum:
          - USER
          - ADMIN
    LoginDto:
      type: object
      properties:
        username:
          maxLength: 32
          minLength: 4
          type: string
          description: логин
        password:
          maxLength: 16
          minLength: 8
          type: string
          description: пароль
...
    UpdateUserDto:
      type: object
      properties:
        firstName:
          maxLength: 10
          minLength: 3
          type: string
          description: имя пользователя
        lastName:
          maxLength: 10
          minLength: 3
          type: string
          description: фамилия пользователя
        phone:
          pattern: "\\+7\\s?\\(?\\d{3}\\)?\\s?\\d{3}-?\\d{2}-?\\d{2}"
          type: string
          description: телефон пользователя
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: id пользователя
          format: int32
          readOnly: true
        email:
          type: string
          description: логин пользователя
        firstName:
          type: string
          description: имя пользователя
        lastName:
          type: string
          description: фамилия пользователя
        phone:
          type: string
          description: телефон пользователя
        image:
          type: string
          description: ссылка на аватар пользователя
        role:
          type: string
          description: роль пользователя
          enum:
          - USER
          - ADMIN
...