# src/main/java/ru/skypro/homework/dto/LoginDto.java
package ru.skypro.homework.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import javax.validation.constraints.Size;

@Data
public class LoginDto {
    @Schema(description = "логин")
    @Size(min = 4, max = 32)
    private String username;

    @Schema(description = "пароль")
    @Size(min = 8, max = 16)
    private String password;


}